Rename quarkus-hibernate-search-backend-elasticsearch to quarkus-hibernate-search-backend-elasticsearch-common
