+++ bom/application/pom.xml
@@ -1377,12 +1377,12 @@
             </dependency>
             <dependency>
                 <groupId>io.quarkus</groupId>
-                <artifactId>quarkus-hibernate-search-backend-elasticsearch</artifactId>
+                <artifactId>quarkus-hibernate-search-backend-elasticsearch-common</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
                 <groupId>io.quarkus</groupId>
-                <artifactId>quarkus-hibernate-search-backend-elasticsearch-deployment</artifactId>
+                <artifactId>quarkus-hibernate-search-backend-elasticsearch-common-deployment</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
