+++ docs/pom.xml
@@ -967,7 +967,7 @@
         </dependency>
         <dependency>
             <groupId>io.quarkus</groupId>
-            <artifactId>quarkus-hibernate-search-backend-elasticsearch-deployment</artifactId>
+            <artifactId>quarkus-hibernate-search-backend-elasticsearch-common-deployment</artifactId>
             <version>${project.version}</version>
             <type>pom</type>
             <scope>test</scope>
