+++ extensions/hibernate-search-backend-elasticsearch-common/runtime/src/main/java/io/quarkus/hibernate/search/backend/elasticsearch/common/runtime/HibernateSearchConfigUtil.java
@@ -1,17 +1,16 @@
-package io.quarkus.hibernate.search.orm.elasticsearch.runtime;
+package io.quarkus.hibernate.search.backend.elasticsearch.common.runtime;
 
+import java.util.LinkedHashSet;
+import java.util.Map;
 import java.util.Optional;
 import java.util.OptionalInt;
+import java.util.Set;
 import java.util.function.BiConsumer;
 import java.util.function.Function;
 
 import org.hibernate.search.engine.cfg.BackendSettings;
 import org.hibernate.search.engine.cfg.IndexSettings;
 
-/**
- * @deprecated Use {@link io.quarkus.hibernate.search.backend.elasticsearch.common.runtime.HibernateSearchConfigUtil} instead.
- */
-@Deprecated
 public class HibernateSearchConfigUtil {
 
     public static <T> void addConfig(BiConsumer<String, Object> propertyCollector, String configPath, T value) {
@@ -69,5 +68,16 @@ public class HibernateSearchConfigUtil {
                 addBackendConfig(propertyCollector, backendName, configPath, getValue.apply(value));
             }
         }
+    }
+
+    public static void mergeInto(Map<String, Set<String>> target, Map<String, Set<String>> source) {
+        for (Map.Entry<String, Set<String>> entry : source.entrySet()) {
+            mergeInto(target, entry.getKey(), entry.getValue());
+        }
+    }
+
+    public static void mergeInto(Map<String, Set<String>> target, String key, Set<String> values) {
+        target.computeIfAbsent(key, ignored -> new LinkedHashSet<>())
+                .addAll(values);
     }
 }
