+++ extensions/hibernate-search-orm-elasticsearch/runtime/src/main/java/io/quarkus/hibernate/search/orm/elasticsearch/runtime/HibernateSearchOrmElasticsearchMapperContext.java
@@ -7,7 +7,7 @@ import java.util.Set;
 
 import org.hibernate.search.engine.environment.bean.BeanReference;
 
-import io.quarkus.hibernate.search.backend.elasticsearch.runtime.MapperContext;
+import io.quarkus.hibernate.search.backend.elasticsearch.common.runtime.MapperContext;
 import io.quarkus.hibernate.search.orm.elasticsearch.runtime.bean.HibernateSearchBeanUtil;
 import io.quarkus.runtime.annotations.RecordableConstructor;
