+++ extensions/hibernate-search-standalone-elasticsearch/deployment/pom.xml
@@ -23,7 +23,7 @@
         </dependency>
         <dependency>
             <groupId>io.quarkus</groupId>
-            <artifactId>quarkus-hibernate-search-backend-elasticsearch-deployment</artifactId>
+            <artifactId>quarkus-hibernate-search-backend-elasticsearch-common-deployment</artifactId>
         </dependency>
         <dependency>
             <groupId>io.quarkus</groupId>
