+++ extensions/pom.xml
@@ -99,7 +99,7 @@
         <module>hibernate-reactive</module>
         <module>hibernate-validator</module>
         <module>panache</module>
-        <module>hibernate-search-backend-elasticsearch</module>
+        <module>hibernate-search-backend-elasticsearch-common</module>
         <module>hibernate-search-orm-elasticsearch</module>
         <module>hibernate-search-orm-outbox-polling</module>
         <module>hibernate-search-standalone-elasticsearch</module>
